+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,12 @@
 # babel-preset-dlight
 
+## 1.0.0-alpha.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/view-generator@1.0.0-alpha.13
+
 ## 1.0.0-alpha.14
 
 ### Patch Changes