+++ packages/tools/transpiler-standalone/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/transpiler-standalone
 
+## 1.0.0-alpha.15
+
+### Patch Changes
+
+- babel-preset-dlight@1.0.0-alpha.15
+
 ## 1.0.0-alpha.14
 
 ### Patch Changes